After completing Solution/Comment Editing and Deleting along with Sorting functionalities 😉😉 (Anuj OP)

+++ components/blogs/Comment.jsx
@@ -1,4 +1,3 @@
-
 import React from "react";
 import Link from "next/link";
 import { useState } from "react";
@@ -10,10 +9,10 @@ import { useRouter } from "next/router";
 import ReactLinkify from "react-linkify";
 import { MoreVertOutlined } from "@material-ui/icons";
 import { IconButton } from "@material-ui/core";
-import Menu from '@material-ui/core/Menu';
-import MenuItem from '@material-ui/core/MenuItem';
-import EditIcon from '@material-ui/icons/Edit';
-import DeleteIcon from '@material-ui/icons/Delete';
+import Menu from "@material-ui/core/Menu";
+import MenuItem from "@material-ui/core/MenuItem";
+import EditIcon from "@material-ui/icons/Edit";
+import DeleteIcon from "@material-ui/icons/Delete";
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
 import DialogContent from "@material-ui/core/DialogContent";
@@ -23,59 +22,64 @@ import CreateIcon from "@material-ui/icons/Create";
 import CircularProgress from "@material-ui/core/CircularProgress";
 import queryFormStyles from "../../styles/components/queries/QueryForm.module.css";
 import Button from "@material-ui/core/Button";
-import Snackbar from '@material-ui/core/Snackbar';
-import MuiAlert from '@material-ui/lab/Alert';
-import { makeStyles } from '@material-ui/core/styles';
+import MuiAlert from "@material-ui/lab/Alert";
+import { makeStyles } from "@material-ui/core/styles";
 import { useStateValue } from "../../context/StateProvider";
+
 function Alert(props) {
   return <MuiAlert elevation={6} variant="filled" {...props} />;
 }
-const Comment = ({ id, by, comment, timestamp,blogBy}) => {
+
+const Comment = ({ id, by, comment, timestamp, blogBy }) => {
   const router = useRouter();
   const [{ user }, dispatch] = useStateValue();
   //const [user, setUser] = useState({});
   const [openMoreVertOutlined, setOpenMoreVertOutlined] = useState(false);
   const [userData, setUserData] = useState({});
   const [anchorEl, setAnchorEl] = React.useState(null);
-  const[editDialogOpen,setEditDialogOpen] = useState(false);
-  const[editedComment,setEditedComment] = useState(comment);
+  const [editDialogOpen, setEditDialogOpen] = useState(false);
+  const [editedComment, setEditedComment] = useState(comment);
   const [posting, setPosting] = useState(false);
+
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
   };
   const handleClose = () => {
     setAnchorEl(null);
   };
+
   const editComment = () => {
     if (editedComment.trim().length < 3) {
       alert("Your comment must be at least 3 characters long.");
+    } else {
+      setEditDialogOpen(true);
+      setPosting(true);
+      db.collection("Blogs")
+        .doc(router.query.blogID)
+        .collection("Comments")
+        .doc(id)
+        .update({
+          comment: editedComment,
+        })
+        .then(() => {
+          setPosting(false);
+          setEditDialogOpen(false);
+          handleClose();
+        });
     }
-    else {
-    setEditDialogOpen(true);
-    setPosting(true);
-    db.collection("Blogs")
-            .doc(router.query.blogID)
-            .collection("Comments")
-            .doc(id)
-            .update({
-              comment : editedComment,
-            })
-            .then(() => {
-              setPosting(false);
-              setEditDialogOpen(false);
-            });
-    }
+  };
 
+  const deleteComment = () => {
+    if (confirm("Are you sure to delete this comment?")) {
+      db.collection("Blogs")
+        .doc(router.query.blogID)
+        .collection("Comments")
+        .doc(id)
+        .delete();
+    }
   };
-  const deleteComment = () =>{
-    db.collection("Blogs")
-            .doc(router.query.blogID)
-            .collection("Comments")
-            .doc(id)
-            .delete();             
-  };
+
   useEffect(() => {
-
     db.collection("Users")
       .doc(by)
       .get()
@@ -106,49 +110,59 @@ const Comment = ({ id, by, comment, timestamp,blogBy}) => {
             </a>
           )}
         >
-          <p className={commentStyles.commentText}>{comment}</p>
+          <pre className={commentStyles.commentText}>{comment}</pre>
         </ReactLinkify>
       </div>
-     
-      <div style={{ display: "flex", flexDirection: "column" }}>
-      { (user?.id === by || user?.id === blogBy) ? 
-        (
-          <div>
-      <MoreVertOutlined onClick = {handleClick} />
-    
-      <Menu
-        id="simple-menu"
-        anchorEl={anchorEl}
-        keepMounted
-        open={Boolean(anchorEl)}
-        onClose={handleClose}
-        style={{
-          display: "flex",
-          justifyContent: "space-between",
-          alignItems: "center",
-        }}
-      >
-        {(user?.id === by ) ? (
-          <div>
-              <MenuItem onClick = {() => setEditDialogOpen(true)} style={{ color: "grey" }}><EditIcon />Edit
-              </MenuItem>
-              <MenuItem onClick={deleteComment} style={{ color: "grey" }}><DeleteIcon/>Delete </MenuItem>
-        
-            </div>
+      {user?.id === by && (
+        <>
+          <IconButton onClick={handleClick}>
+            <MoreVertOutlined />
+          </IconButton>
 
-        ):(
-          <div>
-             
-              <MenuItem onClick={deleteComment} style={{ color: "grey" }}><DeleteIcon/>Delete</MenuItem>
-            </div>
-        )}
-      </Menu>
-      </div>
-        ):(
-          <div>
-  
-      </div>
-        )}
+          <Menu
+            id="simple-menu"
+            anchorEl={anchorEl}
+            keepMounted
+            open={Boolean(anchorEl)}
+            onClose={handleClose}
+            style={{
+              display: "flex",
+              justifyContent: "space-between",
+              alignItems: "center",
+            }}
+          >
+            <MenuItem
+              onClick={() => setEditDialogOpen(true)}
+              style={{
+                color: "grey",
+                display: "flex",
+                justifyContent: "space-between",
+                alignItems: "center",
+              }}
+            >
+              Edit
+              <EditIcon fontSize="small" style={{ marginLeft: 21 }} />
+            </MenuItem>
+            <MenuItem
+              onClick={deleteComment}
+              style={{
+                color: "grey",
+                display: "flex",
+                justifyContent: "space-between",
+                alignItems: "center",
+              }}
+            >
+              Delete
+              <DeleteIcon fontSize="small" style={{ marginLeft: 21 }} />
+            </MenuItem>
+          </Menu>
+        </>
+      )}
+      {user?.id !== by && user?.id === blogBy && (
+        <IconButton onClick={deleteComment}>
+          <DeleteIcon />
+        </IconButton>
+      )}
       <Dialog
         open={editDialogOpen}
         onClose={() => handleClose(false)}
@@ -156,49 +170,43 @@ const Comment = ({ id, by, comment, timestamp,blogBy}) => {
         fullWidth={true}
       >
         <DialogTitle id="form-dialog-title">Edit Comment</DialogTitle>
-          <DialogContent>
-            <div className={queryFormStyles.queryInputDiv}>
-              <CreateIcon
-                fontSize="small"
-                style={{ marginLeft: 4, color: "grey" }}
-              />
-                <TextareaAutosize
-                      placeholder={`Enter your Query`}
-                      autoFocus
-                      id="query-input"
-                      type="text"
-                      className={queryFormStyles.queryInput}
-                      onChange={(e) => setEditedComment(e.target.value)}
-                      value={editedComment}
-                    />
-                    </div>
-                    </DialogContent>
-                    <DialogActions>
-                    <Button onClick={() => setEditDialogOpen(false)} color="primary">
-                      Cancel
-                    </Button>
-                    {posting ? (
-                      <div className="progress-div">
-                        <CircularProgress size={24} style={{ color: "black" }} />
-                      </div>
-                    ) : (
-                      <Button onClick={editComment} color="primary">
-                        Update
-                      </Button>
-                    )}
-                  </DialogActions>
-                    </Dialog>
-       
-        <IconButton>
-        
-        </IconButton>
-      </div>
+        <DialogContent>
+          <div className={queryFormStyles.queryInputDiv}>
+            <CreateIcon
+              fontSize="small"
+              style={{ marginLeft: 4, color: "grey" }}
+            />
+            <TextareaAutosize
+              placeholder={`Enter your Query`}
+              autoFocus
+              id="query-input"
+              type="text"
+              className={queryFormStyles.queryInput}
+              onChange={(e) => setEditedComment(e.target.value)}
+              value={editedComment}
+            />
+          </div>
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={() => setEditDialogOpen(false)} color="primary">
+            Cancel
+          </Button>
+          {posting ? (
+            <div className="progress-div">
+              <CircularProgress size={24} style={{ color: "black" }} />
+            </div>
+          ) : (
+            <Button onClick={editComment} color="primary">
+              Update
+            </Button>
+          )}
+        </DialogActions>
+      </Dialog>
     </div>
   );
 };
 
 export default Comment;
-
 
 // import React from "react";
 // import Link from "next/link";
